refactor	    : Modify the logic for displaying addBookmarkView
- ShareExtension시 URLDataManager을 통해 addBookmarkView를 표시도록 수정
- 앱 꺼짐 상태에서 ShareExtension시 launchscreen 추가
